requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
email-validator==2.3.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0